+++ edb/server/protocol/auth_ext/errors.py
@@ -133,3 +133,23 @@ class UserAlreadyRegistered(AuthExtError):
 
     def __str__(self) -> str:
         return self.description
+
+
+class OAuthProviderFailure(AuthExtError):
+    """OAuth Provider returned a non-success for some part of the flow"""
+
+    def __init__(
+        self,
+        description: str,
+    ):
+        self.description = description
+
+    def __repr__(self):
+        return (
+            f"{self.__class__.__name__}("
+            f"description={self.description!r}"
+            ")"
+        )
+
+    def __str__(self) -> str:
+        return self.description
